pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zng-simulator"
version = "0.1.0"
description = "Commercial 2W Battery Swap Network (BSN) Digital Twin & Financial Simulator"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "streamlit>=1.30,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0,<9.0",
    "pytest-cov>=4.0,<6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/zng_simulator"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
